requirements.txt
pdf2image
PyPDF2
python-docx
Pillow
svgwrite
pdf2docx
img2pdf
PySimpleGUI